# Fix Qwen3-235B A22B with new version of vllm